// src/types/enums.ts
export enum UserRole {
  ADMIN = "admin",
  TEACHER = "teacher",
  STUDENT = "student",
  PARENT = "parent",
}

export enum AttendanceStatusEnum {
  PRESENT = "present",
  ABSENT = "absent",
  LATE = "late",
}

export enum ExamTypeEnum {
  UNIT_TEST = "Unit Test",
  MID_TERM = "Mid Term",
  FINAL = "Final",
  ASSIGNMENT = "Assignment",
  QUIZ = "Quiz",
}

export enum GenderEnum {
  MALE = "Male",
  FEMALE = "Female",
  OTHER = "Other",
}